had to do some refactoring to get the client and server running properly again

+++ GrpcGreeter/Helpers/FakerGenerator.cs
@@ -10,17 +10,17 @@ public static class FakerGenerator
 {
     //there are some redundant Func constructors in here, because sometimes the IDE somtimes likes to display
     //errors that aren't really errors when the RuleFor method is used
-    public static Faker<GameInsertData> GameFaker(List<uint> publisherIds, List<uint> developerIds)
+    public static Faker<GameInsertData> GameFaker(List<int> publisherIds, List<int> developerIds)
     {
         return new Faker<GameInsertData>()
             .RuleFor(g => g.Name, f => f.Commerce.ProductName())
-            .RuleFor(g => g.ReleaseDate, new Func<Faker, uint>(f => 
-                (uint)f.Random.Int(SeedDataConfig.MinReleaseDate, SeedDataConfig.MaxReleaseDate)))
-            .RuleFor(g => g.PublisherId, new Func<Faker, uint>(f => f.PickRandom(publisherIds)))
-            .RuleFor(g => g.DeveloperId, new Func<Faker, uint>(f => f.PickRandom(developerIds)));
+            .RuleFor(g => g.ReleaseDate, new Func<Faker, int>(f => 
+                f.Random.Int(SeedDataConfig.MinReleaseDate, SeedDataConfig.MaxReleaseDate)))
+            .RuleFor(g => g.PublisherId, new Func<Faker, int>(f => f.PickRandom(publisherIds)))
+            .RuleFor(g => g.DeveloperId, new Func<Faker, int>(f => f.PickRandom(developerIds)));
     }
 
-    public static Faker<GameRatingUpsertData> GameRatingUpsertFaker(List<uint> gameIds)
+    public static Faker<GameRatingUpsertData> GameRatingUpsertFaker(List<int> gameIds)
     {
         return new Faker<GameRatingUpsertData>()
             .RuleFor(r => r.GameId, (new Func<Faker, int>(f => (int)f.PickRandom(gameIds))))
@@ -28,7 +28,7 @@ public static class FakerGenerator
             .RuleFor(r => r.Rating, (f => (int)f.Random.Int(SeedDataConfig.MinRating, SeedDataConfig.MaxRating)));
     }
     
-    public static Faker<GameRelation> SemiDeterministicGameRelationFaker(List<uint> gameIds, List<uint> relatedTableIds)
+    public static Faker<GameRelation> SemiDeterministicGameRelationFaker(List<int> gameIds, List<int> relatedTableIds)
     {
         Console.WriteLine($"game ids length: {gameIds.Count}");
         var gameId = 0;
@@ -37,7 +37,7 @@ public static class FakerGenerator
             .RuleFor(gp => gp.RelatedTableId, f => f.PickRandom(relatedTableIds));
     }
     
-    public static Faker<GameRelation> RandomPlatformRelationFaker(List<uint> gameIds, List<uint> relatedTableIds)
+    public static Faker<GameRelation> RandomPlatformRelationFaker(List<int> gameIds, List<int> relatedTableIds)
     {
         return new Faker<GameRelation>()
             .RuleFor(gp => gp.GameId, f => f.PickRandom(gameIds))
